scraper/requirements.txt
# Core dependencies
firecrawl-py==2.12.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1

# Utilities
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
click==8.1.7

# Data processing
pandas==2.1.4
beautifulsoup4==4.12.2

# Monitoring & Logging
structlog==23.2.0
rich==13.7.0

# Development
pytest==7.4.3